Add basic user permissions.

+++ posts/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 
 from .forms import PostForm
@@ -8,6 +8,9 @@ from .models import Post
 
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        return Http404
+
     form = PostForm(request.POST or None, request.FILES or None)
 
     if form.is_valid():
@@ -49,6 +52,9 @@ def post_list(request):
 
 
 def post_update(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        return Http404
+
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 
@@ -67,6 +73,9 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        return Http404
+        
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, 'Post Deleted Successfully', extra_tags='some-tag')
